+++ CHANGELOG.md
@@ -10,10 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Errors now contain the traceback as well as the error message in the result object.
+- Added test for `_post_process` in `tests/covalent_dispatcher_tests/_core/execution_test.py`.
 
 ### Changed
 
 - Post processing logic in `electron` and dispatcher now relies on the order of execution in the transport graph rather than node's function names to allow for a more reliable pairing of nodes and their outputs.
+
+- Renamed `init_test.py` in `tests/covalent_dispatcher_tests/_core/` to `execution_test.py`.
 
 ### Removed